requirements.txt
google-api-core~=2.11
google-api-python-client~=2.7
google-auth~=2.16.0
google-auth-httplib2~=0.1
google-auth-oauthlib~=0.8
googleapis-common-protos~=1.58
openai~=0.26
python-dateutil~=2.8
pytz~=2022.7.1
Django==4.1.6
django-extensions==3.2.1
flask==2.2.2
tweepy==4.12.1
pyyaml==6.0
python-dotenv